fix(cli): Address SonarCloud quality gate
